# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Get Kibble",
        "version": "0.0.1"
    },
    "paths": {
        "/api/auth/password/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Password forget",
                "operationId": "App\\Http\\Controllers\\API\\Swagger\\AuthSwagger::forgotPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@escola-lms.com"
                                    },
                                    "return_url": {
                                        "type": "string",
                                        "example": "https://getkibble.com/password-reset"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/auth/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Password reset",
                "operationId": "App\\Http\\Controllers\\API\\Swagger\\AuthSwagger::resetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@escola-lms.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "hFD4UpgzTMNBeCGyfUrbvHtDb2yzSnuE"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "zaq1@WSX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/auth/social/{provider}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Redirect to social login",
                "operationId": "App\\Http\\Controllers\\API\\Swagger\\AuthSwagger::socialRedirect",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "facebook",
                                "google"
                            ]
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/auth/social/{provider}/callback": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Redirect after sucessfull login",
                "operationId": "App\\Http\\Controllers\\API\\Swagger\\AuthSwagger::socialCallback",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "facebook",
                                "google"
                            ]
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/auth/email/resend": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend email verification",
                "operationId": "App\\Http\\Controllers\\API\\Swagger\\AuthSwagger::resendEmailVerification",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@escola-lms.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "User login",
                "operationId": "App\\Http\\Controllers\\API\\Swagger\\LoginSwagger::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}